refactor: util package로 상수 관리 클래스 이동
